add: routing for recorder
